Controller and Services rafactoring.

+++ Services/API/BlogSite.Business/Concrete/CommentService.cs
@@ -65,17 +65,17 @@ namespace BlogSite.Business.Concrete
             Comment comment = _mapper.Map<Comment>(entityVM);
             comment.Id = Guid.NewGuid();
             comment.CreateTime = DateTime.Now;
-            //comment.UserId = Guid.Parse(_authService.GetCurrentUserId());
+            comment.UserId = Guid.Parse(_authService.GetCurrentUserId());
             var res = await _commentRepository.CreateAsync(comment);
             return res;
         }
 
         public async Task<bool> DeleteAsync(Guid id)
         {
-            //var check = await _commentRepository.GetByIdAsync(id);
-            //var userAuth = Guid.Parse(_authService.GetCurrentUserId());
-            //if(check.UserId == userAuth)
-            //{
+            var check = await _commentRepository.GetByIdAsync(id);
+            var userAuth = Guid.Parse(_authService.GetCurrentUserId());
+            if (check.UserId == userAuth)
+            {
                 var res = await _commentRepository.DeleteAsync(id);
 
                 if (res == true)
@@ -83,26 +83,26 @@ namespace BlogSite.Business.Concrete
                     return true;
                 }
                 return false;
-            //}
-            //return false;
+            }
+            return false;
         }
         public async Task<bool> UpdateAsync(IVM<Comment> entityVM, Guid id)
         {
             Comment comment = _mapper.Map<Comment>(entityVM);
             var check = await _commentRepository.GetByIdAsync(id);
-            //var userAuth = Guid.Parse(_authService.GetCurrentUserId());
+            var userAuth = Guid.Parse(_authService.GetCurrentUserId());
             comment.Id = id;
-            //comment.UserId = userAuth;
-            //if(check.UserId == userAuth)
-            //{
+            comment.UserId = userAuth;
+            if (check.UserId == userAuth)
+            {
                 var res = await _commentRepository.UpdateAsync(comment);
                 if (res == true)
                 {
                     return true;
                 }
                 return false;
-            //}
-            //return false;
+            }
+            return false;
 
         }
