+++ Services/API/BlogSite.Business/Concrete/PostService.cs
@@ -62,7 +62,7 @@ namespace BlogSite.Business.Concrete
             ValidationTool.Validate(new PostValidator(), entityVM);
             Post post = _mapper.Map<Post>(entityVM);
             post.Id = Guid.NewGuid();
-            //post.UserId = Guid.Parse(_authService.GetCurrentUserId());
+            post.UserId = Guid.Parse(_authService.GetCurrentUserId());
             post.CreatedDate = DateTime.Now;
             var res = await _postRepository.CreateAsync(post);
             return res;
@@ -70,33 +70,33 @@ namespace BlogSite.Business.Concrete
 
         public async Task<bool> DeleteAsync(Guid id)
         {
-            //var check = await _postRepository.GetByIdAsync(id);
-            //var userAuth =  Guid.Parse(_authService.GetCurrentUserId());
-            //if(check.UserId == userAuth)
-            //{
+            var check = await _postRepository.GetByIdAsync(id);
+            var userAuth = Guid.Parse(_authService.GetCurrentUserId());
+            if (check.UserId == userAuth)
+            {
                 var res = await _postRepository.DeleteAsync(id);
                 return true;
-            //}
-            //return false;
+            }
+            return false;
         }
 
         public async Task<bool> UpdateAsync(IVM<Post> entityVM, Guid id)
         {
             Post post = _mapper.Map<Post>(entityVM);
             var check = await _postRepository.GetByIdAsync(id);
-//            var userAuth = Guid.Parse(_authService.GetCurrentUserId());
+            var userAuth = Guid.Parse(_authService.GetCurrentUserId());
             post.Id = id;
-            //post.UserId = userAuth;
-                //if (check.UserId == userAuth)
-                //{
-                    var res = await _postRepository.UpdateAsync(post);
+            post.UserId = userAuth;
+            if (check.UserId == userAuth)
+            {
+                var res = await _postRepository.UpdateAsync(post);
                 if (res == true)
                 {
                     return true;
                 }
                 return false;
-            //}
-            //return false;
+            }
+            return false;
         }
 
     }
